Fixed the issue:
The flickering was caused by not refering the object constant buffer using the frame index resource.

When incrementing through update the for loop that iterates updating through each object also determined the object_cb data, this would fail if we added more objects since there can only be three constant buffer objects (one for each frame).

By using frame index, we correctly retrieve the object_cb for the current frame, then in the for loop we update each object in the cb.